refactor: small refactoring of code
