+++ src/App.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import { BrowserRouter as Router } from "react-router-dom";
-import ErrorBoundary from "../components/error-boundary-component/error-boundary-component";
-import AppRoutes from "../components/app-routes/app-routes";
 import "./App.css";
-import { ThemeContext } from "../contexts/themeContext";
+import AppRoutes from "./components/app-routes/app-routes";
+import ErrorBoundary from "./components/error-boundary-component/error-boundary-component";
+import { ThemeContext } from "./contexts/themeContext";
 
 const App: React.FC = () => {
   const [darkTheme, setDarkTheme] = useState<boolean>(false);
