+++ src/main.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./app/App.tsx";
+import App from "./App.tsx";
 import "./index.css";
 import { Provider } from "react-redux";
 import store from "./app/store.ts";
